feat : 유저 모델, 스키마 구현
